explicitly made all fprintf unsinged ints type cast to unsigned long to aid portability between 32bit and 64bit architectures

+++ design/simplicialStructures/diSiTTCSpec.c
@@ -18,15 +18,15 @@ int main() {
   diSiTT_init(disitt1);
 
   fprintf(stdout, "\n>>>diSiTTCSpec\n\n");
-  fprintf(stdout, "size of             int = %u\n", sizeof(int));
-  fprintf(stdout, "size of     dimension_t = %u\n", sizeof(dimension_t));
-  fprintf(stdout, "size of      simplex_id = %u\n", sizeof(simplex_id));
-  fprintf(stdout, "size of           void* = %u\n", sizeof(void*));
-  fprintf(stdout, "size of          DiSiTT = %u\n", sizeof(DiSiTT));
-  fprintf(stdout, "size of    DiSimplexObj = %u",   sizeof(DiSimplexObj));
+  fprintf(stdout, "size of             int = %lu\n", (unsigned long)sizeof(int));
+  fprintf(stdout, "size of     dimension_t = %lu\n", (unsigned long)sizeof(dimension_t));
+  fprintf(stdout, "size of      simplex_id = %lu\n", (unsigned long)sizeof(simplex_id));
+  fprintf(stdout, "size of           void* = %lu\n", (unsigned long)sizeof(void*));
+  fprintf(stdout, "size of          DiSiTT = %lu\n", (unsigned long)sizeof(DiSiTT));
+  fprintf(stdout, "size of    DiSimplexObj = %lu",   (unsigned long)sizeof(DiSimplexObj));
   fprintf(stdout, "   + (dimension + 1)*sizeof(int)\n");
-  fprintf(stdout, "size of DiSimplexObj(1) = %u\n",  DiSimplexObjSize(1));
-  fprintf(stdout, "size of    DiSimplexRef = %u\n", sizeof(DiSimplexRef));
+  fprintf(stdout, "size of DiSimplexObj(1) = %lu\n",  (unsigned long)DiSimplexObjSize(1));
+  fprintf(stdout, "size of    DiSimplexRef = %lu\n", (unsigned long)sizeof(DiSimplexRef));
   fprintf(stdout, "\n");
 
   describe("DiSiTT engine", ^{
